Resize/style/move some widget stuff
