+++ src/components/Button/button.scss
@@ -90,6 +90,14 @@
 }
 
 /*
+ * Slim size
+ */
+.size-slim {
+  padding: 1em 1.15em;
+}
+
+
+/*
  * Regular size
  */
 .size-regular {
