+++ src/components/PetitionStats/petition-stats.scss
@@ -34,7 +34,7 @@
   }
 
   @include media('>text-boost') {
-    font-size: 24px;
+    font-size: 1.5em;
   }
 }
