--- src/components/PetitionWidget/petition-widget.scss
@@ -1,7 +1,0 @@
-@import 'globals';
-
-.root {
-  max-width: 800px;
-  padding: 0.5em 0;
-  background-color: color('white');
-}
