--- src/components/PetitionWidgetHeader/petition-widget-header.scss
@@ -1,24 +1,0 @@
-@import 'globals';
-
-.tags {
-  position: absolute;
-  top: 0;
-  left: 0;
-
-  @include media('>800px') {
-    margin-top: 2.5%;
-  }
-}
-
-.info {
-  list-style: none;
-  font-size: 16px;
-
-  @include media('<=800px') {
-    margin-top: 20px;
-  }
-
-  @include media('>800px') {
-    margin-top: 25px;
-  }
-}
