+++ src/routers/widget.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Route, Router, browserHistory } from 'react-router';
 import logPageview from 'helpers/logPageview';
-import WidgetApp from 'containers/widget/WidgetApp';
-import PetitionWidget from 'containers/widget/PetitionWidget';
+import WidgetApp from 'widgets/containers/WidgetApp';
+import PetitionWidget from 'widgets/containers/PetitionWidget';
 
 export default (props = {}) => (
   <Router history={browserHistory} onUpdate={logPageview}>
