+++ src/widgets/components/PetitionWidgetHeader/petition-widget-header.scss
@@ -1,0 +1,26 @@
+@import 'globals';
+
+.tags {
+  position: absolute;
+  top: 0;
+  left: 0;
+
+  @include media('>800px') {
+    margin-top: 2.5%;
+  }
+}
+
+.info {
+  list-style: none;
+  margin-bottom: -15px;
+
+  @include media('<=text-boost') {
+    margin-top: 12px;
+    font-size: 16px;
+  }
+
+  @include media('>text-boost') {
+    margin-top: 20px;
+    font-size: 15px;
+  }
+}
