+++ src/widgets/components/PetitionWidgetTitle/petition-widget-title.scss
@@ -1,9 +1,7 @@
 @import 'globals';
 
-$bgcolor: color('white');
-
 .title {
+  @include responsive-font(5vw, 18px, 30px);
   margin: 0;
   padding: 0;
-  font-size: 22px;
 }
