+++ src/widgets/components/WidgetBranding/index.js
