+++ src/widgets/components/WidgetBranding/widget-branding.scss
@@ -1,7 +1,15 @@
 @import 'globals';
 
 .root {
-  margin-top: 1em;
+  margin-top: 12px;
+
+  @include media('<=text-boost') {
+    font-size: 13px;
+  }
+
+  @include media('>text-boost') {
+    font-size: 15px;
+  }
 }
 
 .link a:link,
