+++ src/widgets/containers/PetitionWidgetHeader.js
@@ -1,18 +1,14 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import PetitionWidgetHeader from 'components/PetitionWidgetHeader';
+import PetitionWidgetHeader from 'widgets/components/PetitionWidgetHeader';
 import getPetitionHeader from 'selectors/petitionHeader';
-
-const PetitionWidgetHeaderContainer = (props) => (
-  <PetitionWidgetHeader {...props} />
-);
 
 const mapStateToProps = ({ petition }) => getPetitionHeader(petition);
 
-PetitionWidgetHeaderContainer.propTypes = {
+PetitionWidgetHeader.propTypes = {
   title: React.PropTypes.string.isRequired
 };
 
 export default connect(
   mapStateToProps
-)(PetitionWidgetHeaderContainer);
+)(PetitionWidgetHeader);
