+++ src/widgets/containers/PetitionWidgetSupportButton.js
@@ -1,16 +1,16 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import WidgetSupportButton from 'components/WidgetSupportButton';
+import PetitionWidgetSupportButton from 'widgets/components/PetitionWidgetSupportButton';
 import getPetition from 'selectors/petition';
 
 const mapStateToProps = ({ petition }) => ({
   petition: getPetition(petition)
 });
 
-WidgetSupportButton.propTypes = {
+PetitionWidgetSupportButton.propTypes = {
   petition: React.PropTypes.object
 };
 
 export default connect(
   mapStateToProps
-)(WidgetSupportButton);
+)(PetitionWidgetSupportButton);
